feat(users routes): Update User's routes; auth for signup and signin, the rest in users.py and all without app_env paramter + show the user's id in the routes
